requirements.txt
asgiref==3.9.1
crispy-bootstrap5==2025.6
Django==5.2.4
django-crispy-forms==2.4
django-debug-toolbar==6.0.0
django-extensions==4.1
django-htmx==1.23.2
psycopg==3.2.9
psycopg-binary==3.2.9
python-dotenv==1.1.1
sqlparse==0.5.3
